move the entry point for project-control.
