+++ src/packages/hub/copy-path.ts
@@ -11,7 +11,7 @@ import { callback2 } from "@cocalc/util/async-utils";
 import * as message from "@cocalc/util/message";
 import { one_result } from "@cocalc/database";
 import { is_valid_uuid_string, to_json } from "@cocalc/util/misc";
-import { ProjectControlFunction } from "@cocalc/hub/servers/project-control";
+import { ProjectControlFunction } from "@cocalc/server/projects/control";
 
 type WhereQueries = ({ [query: string]: string } | string)[];
