+++ src/packages/hub/proxy/handle-upgrade.ts
@@ -6,7 +6,7 @@ import { versionCheckFails } from "./version";
 import { getTarget } from "./target";
 import getLogger from "../logger";
 import { stripBasePath } from "./util";
-import { ProjectControlFunction } from "@cocalc/hub/servers/project-control";
+import { ProjectControlFunction } from "@cocalc/server/projects/control";
 
 const winston = getLogger("proxy: handle-upgrade");
