+++ src/packages/server/projects/control/kubernetes.ts
@@ -26,7 +26,7 @@ import {
   getProject,
 } from "./base";
 
-import getLogger from "@cocalc/hub/logger";
+import getLogger from "@cocalc/backend/logger";
 const winston = getLogger("project-control-kubernetes");
 
 class Project extends BaseProject {
